+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Changelog
 * Add new command :command:`fuc-undetm`.
 * Add new method :meth:`common.plot_annot_matched`.
 * Add new argument ``sheet`` to :command:`fuc-demux` command.
+* Add new class :class:`common.Variant`.
 
 0.19.0 (2021-07-31)
 -------------------